BhashaEngine/requirements.txt
streamlit==1.31.0
torch
transformers
librosa
sentence-transformers
indic-transliteration
sentencepiece
accelerate
langdetect
gTTS
pyttsx3
streamlit-mic-recorder
spacy
soundfile
protobuf
python-docx
openpyxl
python-pptx
